Create Inscrease and Decrease winround and interact

+++ app/src/main/java/hoangloc/playcardapp/Entity/PlayerList.kt
@@ -12,6 +12,6 @@ data class Player (
 
 
     var name: String = "",
-    var points: Long = 0,
+    var points: Int = 0,
     var winround: Int = 0
 )
